% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{fetch.controls}
\alias{fetch.controls}
\title{Get a curated list of control samples for a specific array}
\usage{
fetch.controls(..., db = c("mm", "mda", "giga"))
}
\arguments{
\item{...}{passed on to array-specific methods}

\item{db}{which database to search: must be one of \code{"mm"} (MegaMUGA), \code{"muga"}, (MUGA) or \code{"giga"} (GigaMUGA)}
}
\value{
a \code{data.frame} of control samples
}
\description{
Get a curated list of control samples for a specific array
}

