% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensities.R
\name{fetch.intensities}
\alias{fetch.intensities}
\title{Retrieve genotype calls and hybridization intensities}
\usage{
fetch.intensities(ids, markers = NULL, chr = NULL, start = NULL,
  end = NULL, raw = FALSE, by = c("id", "name"), db = c("mm", "muga",
  "mda", "giga"), verbose = FALSE)
}
\arguments{
\item{ids}{vector of sample IDs (names or internal database IDs)}

\item{markers}{character vector of marker names for which calls and intensities will be returned}

\item{chr}{limit to markers on these chromosome(s)}

\item{start}{limit to markers with genomic position >= \code{start} (requires \code{chr})}

\item{end}{limit to markers with genomic position <= \code{end} (requires \code{chr})}

\item{raw}{logical; if \code{TRUE}, return raw rather than Illumina-normalized intensities}

\item{by}{search for samples by unique numeric ID (\code{"id"}) or by human-readable sample name (\code{"name"})}

\item{db}{which database to search: must be one of \code{"mm"} (MegaMUGA), \code{"muga"}, (MUGA) or \code{"giga"} (GigaMUGA)}

\item{verbose}{logical; if \code{TRUE}, show progress meter and diagnostic messages to console}
}
\value{
a \code{data.table} with one row per sample-marker combination
}
\description{
Retrieve genotype calls and hybridization intensities
}
\details{
Most of our databases are indexed by sample, chromosome, and position but not by marker name: retrieving data
	for specific markers (as opposed to genomic regions) may be very slow.  It is *always* advisable to use numeric sample
	IDs (guaranteed unique within a database) rather than human-readable names (which may be duplicated).
}

