% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{fetch.samples}
\alias{fetch.samples}
\title{Retrieve a list of samples by name or ID}
\usage{
fetch.samples(ids = NULL, flags = NULL, bad.flags = NULL, group = NULL,
  by = c("name", "id"), exact = TRUE, strict.case = FALSE,
  verbose = TRUE, db = c("mm", "muga", "giga"), ...)
}
\arguments{
\item{ids}{a vector of identifiers to search for; "%" is wildcard character}

\item{bad.flags}{characters in \code{flags} column of database which indicaes a failed sample}

\item{group}{limit to samples from a defined group}

\item{by}{search by unique numeric ID (\code{"id"}) or by human-readable sample name (\code{"name"})}

\item{strict.case}{logical; if \code{TRUE}, perform case-sensitive search}

\item{verbose}{show progress meter on console}

\item{db}{which database to search: must be one of \code{"mm"} (MegaMUGA), \code{"muga"}, (MUGA) or \code{"giga"} (GigaMUGA)}

\item{...}{ignored}
}
\value{
a \code{data.frame} with sample names, IDs and other information
}
\description{
Retrieve a list of samples by name or ID
}

