% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensities.R
\name{make.genotypes}
\alias{make.genotypes}
\title{Create an \code{argyle::genotypes} object via a database query}
\usage{
make.genotypes(ids, snps = NULL, by = c("name", "id"),
  keep.intensity = TRUE, rename = TRUE, make.names.markers = FALSE, ...)
}
\arguments{
\item{ids}{vector of sample IDs (names or internal database IDs)}

\item{snps}{a code{dataframe} containing marker map in \code{argyle} format: columns chromosome, marker name, cM position, physical position, allele 1, allele 2}

\item{by}{search for samples by unique numeric ID (\code{"id"}) or by human-readable sample name (\code{"name})}

\item{keep.intensity}{logical; if \code{TRUE}, return both genotype calls and hybridization intensities}

\item{make.names.markers}{logical; if \code{TRUE}, sanitize marker names of some non-alphanumeric characters with \code{make.names()}}

\item{...}{other options passed to \code{fetch.intensities()}}
}
\value{
a \code{argyle::genotypes} object
}
\description{
Create an \code{argyle::genotypes} object via a database query
}

